Start work on native NuGet package generation support
